requirements.txt
fastapi
uvicorn[standard]
tensorflow
Pillow
numpy